Support saving/loading project files with outdated x-var names
After renaming a column, the model function is not updated, but the plot (meta)data is. Saving and loading a project leaves the fit and model in a somewhat broken state. Really make sure that we can save *and* load such a project and restore the state. The user can then simply update the model function and carry on...