Prevent infinite loop trying to lock in variables
When a stored calculated column expression contains column _names_ (and not _labels_), something's wrong. Maybe the expression was written before a column with that name existed. So, we would try to rename the variables in the expression. However, the (not _labels_) part was not checked and so when columns are still called 'col1', and so on, we would try to rename that expression again and again.