Verify fit data is unchanged before saving project
Fit data was only verified when you activated a plot tab. This made it possible to remove enough data to make a fit fail and save the project without the fit being verified. Loading such a project would result in a crash. Now, saving the project will verify the fit and invalidate it when data was removed.